requirements.txt
vizdoom>=1.3.0
gymnasium>=1.3.0
numpy>=2.0
inferactively-pymdp>=1.0.0
scipy>=1.10
matplotlib>=3.7
pygame-ce>=2.5
